Add an alternative keyboard based on libgpiod
RPi.GPIO seems to be incompatible with newer versions of Raspbian, and
the sysfs interface it uses is also deprecated since Linux 4.8. This
commit adds an alternative keyboard implementation based on the more
modern libgpiod.

Note that libgpiod is not Raspberry Pi specific, but works with all GPIO
hardware that the Linux kernel itself supports. The downside of this is
that the "BOARD" pin numbering mode is not supported.