Add a "global" MQTT client
Unlike the MQTT keyboard, which is meant to provide user-configurable
input and output facilities, this client publishes some fixed internal
status information. This allows it to publish much more detailed
information than is possible with a keyboard instance.

Currently published are a button to hangup all calls, one to reload the
configuration file, a sensor with the name of the SIP phone
implementation, and a sensor that shows the current call state (inactive
/ ringing / connected), along with some details about the call.